low level 128-bit submul on bigtorus

+++ fhe/BigTorusPolynomial.cpp
@@ -8,10 +8,8 @@ BigTorusPolynomial::~BigTorusPolynomial() {
 }
 
 void const_poly(BigTorusPolynomial &out, const BigTorusRef &in, UINT64 out_limb_prec) {
-
     copy(out.getAT(0), in, out_limb_prec);
-
-    for (int i = 1; i < out.length; i++) {
+    for (UINT64 i = 1; i < out.length; i++) {
         zero(out.getAT(i));
     }
 }
